tools/requirements.txt
termcolor
boto3
python-dotenv